added rakecounter

+++ Core/Toolbox/UserMapping.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace Rake_Counter.Core.Toolbox
 {
-    internal class UserMapping : Dictionary<string, float>
+    internal class UserMapping : Dictionary<string, double>
     {
-        public void Map(string user, float amount)
+        public void Map(string user, double amount)
         {
             if (!this.ContainsKey(user))
             {
